# src/swagger.yaml
...
swagger: "2.0"
info:
  description: api used in hey-node app
  version: 1.0.0
  title: Hey-node API
  contact:
    email: gorbunovano@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
host: virtserver.swaggerhub.com
basePath: /hey-node/node/1.0.0
...
schemes:
- https
paths:
  /users:
    get:
      tags:
      - users
      description: get list of users
      parameters: []
      responses:
        200:
          description: Successful response
          schema:
            type: array
            items:
              $ref: '#/definitions/User'
          responseSchema:
            type: array
            items:
              $ref: '#/definitions/User'
        500:
          description: Server error
      security:
      - jwt: []
  /users/{userId}:
    delete:
      tags:
      - users
      description: delete user by id
      parameters:
      - name: userId
        in: path
        description: Id of user to delete
        required: true
        type: string
        format: uuid
      responses:
        200:
          description: Successful response
        401:
          description: Not authorized
        500:
          description: Server error
      security:
      - jwt: []
  /auth:
    post:
      tags:
      - users
      description: auth endpoint
      parameters:
      - in: body
        name: credentials
        description: credentials to login
        required: false
        schema:
          $ref: '#/definitions/credentials'
      responses:
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/inline_response_200'
          responseSchema:
            type: object
            properties:
              user:
                $ref: '#/definitions/User'
              token:
                type: string
                example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI1YjM5MmY2ODM3YTUwZDViYTQyNzEwNGYiLCJpYXQiOjE1MzIyNjM5NjksImV4cCI6MTUzMjI2NDk2OX0.Mw0A7JJW2dUCbx__9YjpTZWD40EWI_XXPlamxvmKuyY
        404:
          description: User is not found
        500:
          description: Server error
      security: []
...
securityDefinitions:
  jwt:
    description: Secures endpoint by requiring a valid JWT token
    type: apiKey
    name: Authentication
    in: header
definitions:
  User:
    type: object
    required:
    - id
    - login
    - name
    properties:
      id:
        type: string
        format: uuid
        example: d290f1ee-6c54-4b01-90e6-d701748f0851
      name:
        type: string
        example: Alex
      login:
        type: string
        example: alex
      lastModifiedDate:
        type: string
        example: 2016-08-29T09:12:33.001Z
    example:
      lastModifiedDate: 2016-08-29T09:12:33.001Z
      name: Alex
      id: d290f1ee-6c54-4b01-90e6-d701748f0851
      login: alex
...
  credentials:
    type: object
    required:
    - login
    - password
    properties:
      login:
        type: string
      password:
        type: string
  inline_response_200:
    properties:
      user:
        $ref: '#/definitions/User'
      token:
        type: string
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI1YjM5MmY2ODM3YTUwZDViYTQyNzEwNGYiLCJpYXQiOjE1MzIyNjM5NjksImV4cCI6MTUzMjI2NDk2OX0.Mw0A7JJW2dUCbx__9YjpTZWD40EWI_XXPlamxvmKuyY
    example:
      user:
        lastModifiedDate: 2016-08-29T09:12:33.001Z
        name: Alex
        id: d290f1ee-6c54-4b01-90e6-d701748f0851
        login: alex
      token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI1YjM5MmY2ODM3YTUwZDViYTQyNzEwNGYiLCJpYXQiOjE1MzIyNjM5NjksImV4cCI6MTUzMjI2NDk2OX0.Mw0A7JJW2dUCbx__9YjpTZWD40EWI_XXPlamxvmKuyY
...